refactor: episode controller 분리
