+++ src/main/java/com/example/webnovel/book/ui/BookController.java
@@ -4,13 +4,10 @@ import com.example.webnovel.book.application.BookService;
 import com.example.webnovel.book.domain.book.type.BookStatus;
 import com.example.webnovel.book.dto.BookCreateRequest;
 import com.example.webnovel.book.dto.BookResponse;
-import com.example.webnovel.book.dto.EpisodeCreateRequest;
-import com.example.webnovel.book.dto.EpisodeResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -48,23 +45,5 @@ public class BookController {
     ) {
         Page<BookResponse> result = bookService.getBooks(bookStatus, pageable);
         return ResponseEntity.ok(result);
-    }
-
-    @GetMapping("/{bookId}/episodes/{episodeId}")
-    public ResponseEntity<EpisodeResponse> getEpisode(
-            @PathVariable Long bookId,
-            @PathVariable Long episodeId
-    ) {
-        final EpisodeResponse episode = bookService.getEpisode(bookId, episodeId);
-        return ResponseEntity.ok(episode);
-    }
-
-    @PostMapping("/{bookId}/episodes")
-    public ResponseEntity<EpisodeResponse> addEpisode(
-            @PathVariable Long bookId,
-            @RequestBody @Valid EpisodeCreateRequest request
-    ) {
-        bookService.addEpisode(bookId, request.getTitle(), request.getContent(), request.getTicketPrice());
-        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 }
